Reword Scrutinizer PHP version specification
Scrutinizer started complaining about not finding the version specified.
According to the docs it should infer the version automatically when
omitted, but it doesn't (fell back all the way to PHP 5.4, even though
composer.json specifies at least 7) so I guess that doesn't work too well.

With this fix I hope it takes the latest version that fits the pattern.